Refactor Modern and Legacy CryptoStrategies to share a base class
This simplifies some code sharing, and gives us one spot to validate
RP-Key
