+++ src/remoteplay/crypto/modern.ts
@@ -1,43 +1,15 @@
-import crypto from "crypto";
-import { IRemotePlayCredentials } from "../../credentials/model";
 import { DeviceType } from "../../discovery/model";
-import { CryptoCodec } from "../../socket/crypto-codec";
 
 import { RemotePlayVersion } from "../model";
-import { padBuffer, parseHexBytes } from "../protocol";
+import { BaseCryptoStrategy } from "./base";
 import {
     AERO_KEYS,
     AUTH_NONCE_KEYS,
     AUTH_SEED_KEYS,
     INIT_KEYS,
 } from "./keys";
-import { ICryptoStrategy } from "./model";
 
 const SEED_BYTES_COUNT = 16;
-const PADDING_BYTES = 480;
-const CRYPTO_ALGORITHM = "aes-128-cfb";
-
-const hmacKeys = {
-    [RemotePlayVersion.PS5_1]: "464687b349ca8ce859c5270f5d7a69d6",
-    [RemotePlayVersion.PS4_10]: "20d66f5904ea7c14e557ffc52e488ac8",
-    [RemotePlayVersion.PS4_9]: "ac078883c83a1fe811463af39ee3e377",
-    [RemotePlayVersion.PS4_8]: "ac078883c83a1fe811463af39ee3e377",
-};
-
-// NOTE: public for testing
-export function generateIv(version: RemotePlayVersion, nonce: Buffer, counter: bigint) {
-    const counterBuffer = Buffer.alloc(8, "binary");
-    counterBuffer.writeBigUInt64BE(counter);
-
-    const hmacKey = hmacKeys[version];
-    const hmac = crypto.createHmac("sha256", Buffer.from(hmacKey, "hex"));
-
-    hmac.update(nonce);
-    hmac.update(counterBuffer);
-    const digest = hmac.digest();
-
-    return digest.slice(0, SEED_BYTES_COUNT);
-}
 
 // NOTE: public for testing
 export function generateSeed(
@@ -145,24 +117,23 @@ export function generateAuthSeed(
 /**
  * Handles crypto for PS5s, and PS4s on RemotePlay 10.0+
  */
-export class ModernCryptoStrategy implements ICryptoStrategy {
+export class ModernCryptoStrategy extends BaseCryptoStrategy {
     constructor(
         private readonly deviceType: DeviceType,
-        private readonly version: RemotePlayVersion,
+        version: RemotePlayVersion,
     ) {
+        super(version);
     }
 
-    public createCodecForPin(pin: string, nonce: Buffer) {
-        const pinNumber = parseInt(pin, 10);
-
-        const padding = Buffer.alloc(PADDING_BYTES).fill("A");
-
+    protected generatePinSeed(padding: Buffer, pinNumber: number): Buffer {
         /* eslint-disable no-bitwise */
         const initKeyOff = padding[0x18D] & 0x1F;
         /* eslint-enable no-bitwise */
 
-        const iv = generateIv(this.version, nonce, /* counter = */BigInt(0));
-        const seed = generateSeed(this.deviceType, pinNumber, initKeyOff);
+        return generateSeed(this.deviceType, pinNumber, initKeyOff);
+    }
+
+    protected signPadding(nonce: Buffer, padding: Buffer) {
         const aeropause = generateAeropause(this.deviceType, nonce, padding);
 
         const AEROPAUSE_PART1_DESTINATION = 0xc7;
@@ -170,27 +141,13 @@ export class ModernCryptoStrategy implements ICryptoStrategy {
 
         aeropause.copy(padding, AEROPAUSE_PART1_DESTINATION, 8, 16);
         aeropause.copy(padding, AEROPAUSE_PART2_DESTINATION, 0, 8);
-
-        const codec = new CryptoCodec(iv, seed, CRYPTO_ALGORITHM);
-        return {
-            preface: padding,
-            codec,
-        };
     }
 
-    public createCodecForAuth(
-        creds: IRemotePlayCredentials,
-        serverNonce: Buffer,
-        counter: bigint,
-    ) {
-        // this is known as "morning" to chiaki for some reason
-        const key = padBuffer(parseHexBytes(creds.registration["RP-Key"]));
+    protected generateAuthSeed(key: Buffer, serverNonce: Buffer) {
+        return generateAuthSeed(this.deviceType, key, serverNonce);
+    }
 
-        const nonce = transformServerNonceForAuth(this.deviceType, serverNonce);
-        const seed = generateAuthSeed(this.deviceType, key, serverNonce);
-
-        const iv = generateIv(this.version, nonce, counter);
-
-        return new CryptoCodec(iv, seed, CRYPTO_ALGORITHM);
+    protected transformServerNonceForAuth(serverNonce: Buffer) {
+        return transformServerNonceForAuth(this.deviceType, serverNonce);
     }
 }
